refactor: errors

+++ src/Schemata.Abstractions/Resource/ListResult.cs
@@ -6,7 +6,7 @@ public class ListResult<TSummary> : OperationResult<ListResult<TSummary>>
 {
     public virtual IEnumerable<TSummary>? Entities { get; set; }
 
-    public virtual long? TotalSize { get; set; }
+    public virtual int? TotalSize { get; set; }
 
     public virtual string? NextPageToken { get; set; }
